backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pyotp
passlib[bcrypt]
python-dotenv
pydantic[email]
